# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Denarixx OneEarth API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UnifiedAlertModule =
  (typeof UnifiedAlertModule)[keyof typeof UnifiedAlertModule];

export const UnifiedAlertModule = {
  energy: "energy",
  lifemesh: "lifemesh",
  earthshield: "earthshield",
} as const;

export type UnifiedAlertSeverity =
  (typeof UnifiedAlertSeverity)[keyof typeof UnifiedAlertSeverity];

export const UnifiedAlertSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export type UnifiedAlertStatus =
  (typeof UnifiedAlertStatus)[keyof typeof UnifiedAlertStatus];

export const UnifiedAlertStatus = {
  active: "active",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;

export interface UnifiedAlert {
  id: number;
  title: string;
  module: UnifiedAlertModule;
  severity: UnifiedAlertSeverity;
  location: string;
  status: UnifiedAlertStatus;
  description: string;
  createdAt: string;
}

export interface DashboardStats {
  activeSites: number;
  criticalAlerts: number;
  protectedPeople: number;
  disasterRiskZones: number;
  energyAvailability: number;
  recentAlerts: UnifiedAlert[];
}

export type SiteType = (typeof SiteType)[keyof typeof SiteType];

export const SiteType = {
  village: "village",
  clinic: "clinic",
  school: "school",
  district: "district",
  shelter: "shelter",
} as const;

export type SiteStatus = (typeof SiteStatus)[keyof typeof SiteStatus];

export const SiteStatus = {
  online: "online",
  offline: "offline",
  warning: "warning",
  critical: "critical",
} as const;

export type SiteCurrentRiskLevel =
  (typeof SiteCurrentRiskLevel)[keyof typeof SiteCurrentRiskLevel];

export const SiteCurrentRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Site {
  id: number;
  name: string;
  type: SiteType;
  location: string;
  country: string;
  status: SiteStatus;
  uptime: number;
  powerAvailability: number;
  currentRiskLevel: SiteCurrentRiskLevel;
  population: number;
  latitude: number;
  longitude: number;
  createdAt: string;
}

export type CreateSiteInputType =
  (typeof CreateSiteInputType)[keyof typeof CreateSiteInputType];

export const CreateSiteInputType = {
  village: "village",
  clinic: "clinic",
  school: "school",
  district: "district",
  shelter: "shelter",
} as const;

export interface CreateSiteInput {
  name: string;
  type: CreateSiteInputType;
  location: string;
  country: string;
  latitude: number;
  longitude: number;
  population: number;
}

export type EnergyMetricGridStatus =
  (typeof EnergyMetricGridStatus)[keyof typeof EnergyMetricGridStatus];

export const EnergyMetricGridStatus = {
  stable: "stable",
  unstable: "unstable",
  offline: "offline",
} as const;

export interface EnergyMetric {
  id: number;
  siteId: number;
  siteName: string;
  solarGeneration: number;
  batteryLevel: number;
  communityLoad: number;
  gridStatus: EnergyMetricGridStatus;
  alerts: string[];
  uptime: number;
  recordedAt: string;
}

export interface EnergyChartPoint {
  time: string;
  solar: number;
  battery: number;
  load: number;
}

export type ProtectedPersonCategory =
  (typeof ProtectedPersonCategory)[keyof typeof ProtectedPersonCategory];

export const ProtectedPersonCategory = {
  child: "child",
  elderly: "elderly",
  family: "family",
  vulnerable: "vulnerable",
} as const;

export type ProtectedPersonStatus =
  (typeof ProtectedPersonStatus)[keyof typeof ProtectedPersonStatus];

export const ProtectedPersonStatus = {
  safe: "safe",
  "at-risk": "at-risk",
  emergency: "emergency",
  unknown: "unknown",
} as const;

export interface ProtectedPerson {
  id: number;
  name: string;
  age: number;
  category: ProtectedPersonCategory;
  status: ProtectedPersonStatus;
  lastKnownLocation: string;
  contactName: string;
  contactPhone: string;
  siteId: number;
  siteName: string;
  notes?: string;
  updatedAt: string;
}

export type SafetyIncidentSeverity =
  (typeof SafetyIncidentSeverity)[keyof typeof SafetyIncidentSeverity];

export const SafetyIncidentSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export type SafetyIncidentStatus =
  (typeof SafetyIncidentStatus)[keyof typeof SafetyIncidentStatus];

export const SafetyIncidentStatus = {
  open: "open",
  "in-progress": "in-progress",
  resolved: "resolved",
} as const;

export interface SafetyIncident {
  id: number;
  title: string;
  description: string;
  severity: SafetyIncidentSeverity;
  location: string;
  involvedPersons: string[];
  status: SafetyIncidentStatus;
  responseChain: string[];
  occurredAt: string;
}

export interface SOSInput {
  personId: number;
  location: string;
  message?: string;
}

export interface SOSResponse {
  success: boolean;
  incidentId: number;
  message: string;
  responseTeam: string[];
}

export type DisasterAlertType =
  (typeof DisasterAlertType)[keyof typeof DisasterAlertType];

export const DisasterAlertType = {
  flood: "flood",
  wildfire: "wildfire",
  storm: "storm",
  earthquake: "earthquake",
  infrastructure: "infrastructure",
  drought: "drought",
} as const;

export type DisasterAlertSeverity =
  (typeof DisasterAlertSeverity)[keyof typeof DisasterAlertSeverity];

export const DisasterAlertSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export type DisasterAlertStatus =
  (typeof DisasterAlertStatus)[keyof typeof DisasterAlertStatus];

export const DisasterAlertStatus = {
  active: "active",
  monitoring: "monitoring",
  resolved: "resolved",
} as const;

export interface DisasterAlert {
  id: number;
  title: string;
  type: DisasterAlertType;
  severity: DisasterAlertSeverity;
  region: string;
  country: string;
  description: string;
  affectedPopulation: number;
  status: DisasterAlertStatus;
  issuedAt: string;
}

export type RiskZoneType = (typeof RiskZoneType)[keyof typeof RiskZoneType];

export const RiskZoneType = {
  flood: "flood",
  wildfire: "wildfire",
  storm: "storm",
  earthquake: "earthquake",
  infrastructure: "infrastructure",
} as const;

export type RiskZoneRiskLevel =
  (typeof RiskZoneRiskLevel)[keyof typeof RiskZoneRiskLevel];

export const RiskZoneRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface RiskZone {
  id: number;
  name: string;
  type: RiskZoneType;
  riskLevel: RiskZoneRiskLevel;
  region: string;
  country: string;
  preparednessScore: number;
  latitude: number;
  longitude: number;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  operator: "operator",
  family: "family",
  government: "government",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  organization: string;
  status: UserStatus;
  lastLogin?: string;
  createdAt: string;
}

export type GetEnergyMetricsParams = {
  siteId?: number;
};

export type GetEnergyChartParams = {
  siteId?: number;
};

export type GetUnifiedAlertsParams = {
  module?: GetUnifiedAlertsModule;
  severity?: GetUnifiedAlertsSeverity;
};

export type GetUnifiedAlertsModule =
  (typeof GetUnifiedAlertsModule)[keyof typeof GetUnifiedAlertsModule];

export const GetUnifiedAlertsModule = {
  energy: "energy",
  lifemesh: "lifemesh",
  earthshield: "earthshield",
} as const;

export type GetUnifiedAlertsSeverity =
  (typeof GetUnifiedAlertsSeverity)[keyof typeof GetUnifiedAlertsSeverity];

export const GetUnifiedAlertsSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;