# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Denarixx OneEarth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: getSites
      tags: [sites]
      summary: List all sites
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get a site by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
...
  /energy/chart:
    get:
      operationId: getEnergyChart
      tags: [energy]
      summary: Get energy chart data for a site
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnergyChartPoint"
...
  /lifemesh/sos:
    post:
      operationId: triggerSOS
      tags: [lifemesh]
      summary: Trigger SOS alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SOSInput"
      responses:
        "200":
          description: SOS triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOSResponse"
...
components:
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [village, clinic, school, district, shelter]
        location:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [online, offline, warning, critical]
        uptime:
          type: number
        powerAvailability:
          type: number
        currentRiskLevel:
          type: string
          enum: [low, medium, high, critical]
        population:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - location
        - country
        - status
        - uptime
        - powerAvailability
        - currentRiskLevel
        - population
        - latitude
        - longitude
        - createdAt
    CreateSiteInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [village, clinic, school, district, shelter]
        location:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        population:
          type: integer
      required:
        - name
        - type
        - location
        - country
        - latitude
        - longitude
        - population
...
    EnergyChartPoint:
      type: object
      properties:
        time:
          type: string
        solar:
          type: number
        battery:
          type: number
        load:
          type: number
      required:
        - time
        - solar
        - battery
        - load
...
    SOSInput:
      type: object
      properties:
        personId:
          type: integer
        location:
          type: string
        message:
          type: string
      required:
        - personId
        - location
    SOSResponse:
      type: object
      properties:
        success:
          type: boolean
        incidentId:
          type: integer
        message:
          type: string
        responseTeam:
          type: array
          items:
            type: string
      required:
        - success
        - incidentId
        - message
        - responseTeam
...