# lib/api-zod/src/generated/types/energyChartPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Denarixx OneEarth API
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyChartPoint {
  time: string;
  solar: number;
  battery: number;
  load: number;
}