# lib/api-zod/src/generated/types/sOSInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Denarixx OneEarth API
 * OpenAPI spec version: 0.1.0
 */

export interface SOSInput {
  personId: number;
  location: string;
  message?: string;
}